# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "complain_management_sys",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://127.0.0.1"
    }
  ],
  "paths": {
...
    "/complaint/add_type": {
      "post": {
        "tags": [
          "complaint"
        ],
        "summary": "add_complaint_type",
        "description": "add_complaint_type",
        "operationId": "addComplaintType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "critical"
                  }
                }
              },
              "examples": {
                "add_complaint_type": {
                  "value": {
                    "name": "critical"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New Request",
            "headers": {
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "close"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "120"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 08 Dec 2023 18:06:20 GMT"
                }
              },
              "Server": {
                "schema": {
                  "type": "string",
                  "example": "Werkzeug/3.0.1 Python/3.11.5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "message": {
                      "type": "string",
                      "example": "Complaint type added successfully"
                    },
                    "status_code": {
                      "type": "number",
                      "example": 200
                    }
                  }
                },
                "examples": {
                  "New Request": {
                    "value": {
                      "data": {
                        "id": 1
                      },
                      "errors": [],
                      "message": "Complaint type added successfully",
                      "status_code": 200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/complaint/add": {
      "post": {
        "tags": [
          "complaint"
        ],
        "summary": "add_complaint",
        "description": "add_complaint",
        "operationId": "addComplaint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "complaint_type_id": {
                    "type": "number",
                    "example": 1
                  },
                  "description": {
                    "type": "string",
                    "example": "complaint description"
                  },
                  "title": {
                    "type": "string",
                    "example": "new complaint"
                  }
                }
              },
              "examples": {
                "add_complaint": {
                  "value": {
                    "complaint_type_id": 1,
                    "description": "complaint description",
                    "title": "new complaint"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "add_complaint",
            "headers": {
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "close"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "115"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 08 Dec 2023 19:02:36 GMT"
                }
              },
              "Server": {
                "schema": {
                  "type": "string",
                  "example": "Werkzeug/3.0.1 Python/3.11.5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "message": {
                      "type": "string",
                      "example": "Complaint added successfully"
                    },
                    "status_code": {
                      "type": "number",
                      "example": 200
                    }
                  }
                },
                "examples": {
                  "add_complaint": {
                    "value": {
                      "data": {
                        "id": 1
                      },
                      "errors": [],
                      "message": "Complaint added successfully",
                      "status_code": 200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/complaint/update/14": {
      "put": {
        "tags": [
          "complaint"
        ],
        "summary": "update_complaint",
        "description": "update_complaint",
        "operationId": "updateComplaint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "complaint_type_id": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "complaint title"
                  }
                }
              },
              "examples": {
                "update_complaint": {
                  "value": {
                    "complaint_type_id": 1,
                    "title": "complaint title"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "update_complaint",
            "headers": {
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "close"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "102"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 08 Dec 2023 20:28:12 GMT"
                }
              },
              "Server": {
                "schema": {
                  "type": "string",
                  "example": "Werkzeug/3.0.1 Python/3.11.5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "message": {
                      "type": "string",
                      "example": "Complaint updated successfully"
                    },
                    "status_code": {
                      "type": "number",
                      "example": 200
                    }
                  }
                },
                "examples": {
                  "update_complaint": {
                    "value": {
                      "data": [],
                      "errors": [],
                      "message": "Complaint updated successfully",
                      "status_code": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "update_complaint",
            "headers": {
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "close"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "136"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 08 Dec 2023 20:28:21 GMT"
                }
              },
              "Server": {
                "schema": {
                  "type": "string",
                  "example": "Werkzeug/3.0.1 Python/3.11.5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "No update detect in given data or Invalid complaint id"
                      },
                      "example": [
                        "No update detect in given data or Invalid complaint id"
                      ]
                    },
                    "message": {
                      "type": "string",
                      "example": ""
                    },
                    "status_code": {
                      "type": "number",
                      "example": 400
                    }
                  }
                },
                "examples": {
                  "update_complaint": {
                    "value": {
                      "data": [],
                      "errors": [
                        "No update detect in given data or Invalid complaint id"
                      ],
                      "message": "",
                      "status_code": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/complaint/1/done": {
      "patch": {
        "tags": [
          "complaint"
        ],
        "summary": "done_complaint",
        "description": "done_complaint",
        "operationId": "doneComplaint",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "done_complaint",
            "headers": {
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "close"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "101"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 08 Dec 2023 22:01:00 GMT"
                }
              },
              "Server": {
                "schema": {
                  "type": "string",
                  "example": "Werkzeug/3.0.1 Python/3.11.5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "message": {
                      "type": "string",
                      "example": "Complaint doneed successfully"
                    },
                    "status_code": {
                      "type": "number",
                      "example": 200
                    }
                  }
                },
                "examples": {
                  "done_complaint": {
                    "value": {
                      "data": [],
                      "errors": [],
                      "message": "Complaint doneed successfully",
                      "status_code": 200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}